fix: Fabric8 Kubernetes Client is not deprecated
Fabric8 Kubernetes Client is still alive and kicking!
Version 6.0.0 GA to be released on early July with many
ground-breaking and game changer features.

Eclipse JKube is the successor of Fabric8 Maven Plugin (FMP),
which is deprecated. JKube includes Maven Plugins for
the deployment of Java Applications into K8s, just as FMP.
And also equivalent Gradle plugins (which are new).